+++ autogpt_platform/frontend/src/app/marketplace/search/page.tsx
