+++ autogpt_platform/frontend/src/app/page.tsx
@@ -3,5 +3,5 @@
 import { redirect } from "next/navigation";
 
 export default function Page() {
-  redirect("/store");
+  redirect("/marketplace");
 }
