+++ autogpt_platform/frontend/src/components/agptui/composite/FeaturedCreators.tsx
@@ -24,7 +24,7 @@ export const FeaturedCreators: React.FC<FeaturedCreatorsProps> = ({
   const router = useRouter();
 
   const handleCardClick = (creator: string) => {
-    router.push(`/store/creator/${encodeURIComponent(creator)}`);
+    router.push(`/marketplace/creator/${encodeURIComponent(creator)}`);
   };
 
   // Only show first 4 creators
